+++ DatadogRUM/Sources/RUMMonitor/RUMCommand.swift
@@ -460,14 +460,24 @@ internal protocol RUMResourceCommand: RUMCommand {
 }
 
 /// Tracing information propagated by Tracing to the underlying `URLRequest`. It is passed to the RUM backend
-/// in order to create the APM span. The actual `Span` is not send by the SDK.
+/// in order to create the APM span. The actual `Span` is not sent by the SDK.
 internal struct RUMSpanContext {
     /// The trace ID injected to `URLRequest` that issues RUM resource.
     let traceID: TraceID
     /// The span ID injected to `URLRequest` that issues RUM resource.
     let spanID: SpanID
+    /// The parent span ID injected to `URLRequest` that issues RUM resource.
+    /// This is the ID of the active span if it exists, and Trace feature is enabled.
+    let parentSpanID: SpanID?
     /// The sampling rate applied to the trace (a value between `0.0` and `1.0`).
     let samplingRate: Double
+
+    init(traceID: TraceID, spanID: SpanID, parentSpanID: SpanID?, samplingRate: Double) {
+        self.traceID = traceID
+        self.spanID = spanID
+        self.parentSpanID = parentSpanID
+        self.samplingRate = samplingRate
+    }
 }
 
 internal struct RUMStartResourceCommand: RUMResourceCommand {